ctl: split into multiple modules
ref #10938
